refacto: extract telemetry stuff into its own module
feat: doc new module
